Fix for nosetest import issues
The nosetests were importing the local version of pysatCDF instead of the one in the virtualenv.
The local folder is the first entry in sys.path, so the local pysatCDF folder will hit before the one
installed by setup.py. Using develop ensures that the local copy will have fortran_cdf.so. Previous builds
failed because the file only existed in site-packages and not locally.